regression: add type support in quantified effects
If an effect is of the form all(Var,Type,Effect), then check if the term
is of type Type and only substitute if it is. Thus, the effect is only
applied to a term if the term is known to be of type Type. If a term is
of different type (or a supertype), then the term is not substituted and
thus the term is not removed from the condition to be regressed.