regression: on a cond effect, regress both cases
We need to regress a term on a conditional effect for both cases
(condition true or false). If the condition is false, then no effect is
applied, but we need to add not(Cond) to the regressed term. Similarly,
if the condition is true, regress the effect and add Cond to the
regressed term.