mongodb: allow user configuration for multiple databases
This allows to configure users for multiple databases: Simply specify
the database in your config as follows:
  - {
      user: "alice",
      pwd: "supersecretpassword",
      roles: '[ { role: \"userAdminAnyDatabase\", db: \"admin\" },
               { role: \"readWrite\", db: \"mydatabase\" } ]',
      db: "mydatabase"
    }

All users are stored in the system's user database, but we can have the
same user in different databases. Note that a user's role for a database
gives the user access to the database but does not necessarily add the
user to the database.