kubernetes: change from job queue to job template expansion
Instead of using a job queue that is accessed from the container, create
one job for each task and directly start the worker in the container.
This avoids having to access a queue from within the container, and
makes it more transparent which tasks are currently running.

This creates one job for each planning task instead of one job for all
tasks together. See
http://kubernetes.io/docs/user-guide/jobs/#job-patterns for more
information on job patterns.