identification: directly identify common parameter assignments
Instead of checking the parameter assignment in a separate step,
directly identify frequent parameter assignments while computing
frequent action sequences. This computes a list of frequent action
sequences, and for each sequence a list of frequent parameter
assignments including the count of the assignment.

Note that the parameter assignment is always the most specific
assignment possible, i.e., additional steps are necessary to count the
occurrences of a generalized action sequence (i.e., a sequence where all
parameters are distinct).