planning worker: also write domain and used resources into the database
We need the domain so we can easily query for all solutions of a certain
domain, which is necessary for querying for frequent action sequences.
Also write used resources such as CPU time and memory into the database.
See the python resource module for the details on the resource array.