identification: output of Map and Reduce must have the same type
The output of Map must have the same structure as the output of Reduce.
Reduce is only called if multiple entries exist for the same key; if a
key exists only once then the result of Map is directly written in the
result collection. This means that previously, any sequences that
occurred only once had the wrong structure in the result collection,
leading to many null entries.

Also add a totalCount for each action sequence. The total count is the
sum over all possible parameterizations of one action sequences.