planning worker: write to database whether result can be used for macros
We want to run planners for benchmarks, in which case the results should
not be used for macro generation. Thus, introduce a new database field
'use_for_macros' which is set to true if the result can be used for
macro generation. This must be passed to the worker by the caller.