substitute: also allow string atoms in the terms list
Strings are never treated as atoms (i.e., atom(S) is false for any
string S). If term contains an atomic string (e.g., "a"), then calling
substitute on all subterms results in infinite recursion because "a" =..
["a"] holds and thus we call the same substitution over and over again.
To avoid this, check if the resulting list of subterms is the same as
the term itself. If so, do not call the substitution on the subterms.