generation: add effect tree implementation
Effect trees are used to compute the effect of an action sequence with
conditional effects. As the effect of any action can depend on the
conditional effect of a subsequent action, we can no longer compute the
effect in a linear fashion. Instead, create a tree where each split
represents a conditional effect where the condition was set to
true/false.