regression: fix regression on an empty effects list
If the Effects are empty, then the parameter types can be anything and
don't have to be empty.

This error occurred because of inconsistent use of ParameterTypes. It
was initially intended as a list of lists that describe each parameter
set of every effect separately. Now, we assume that the ParameterTypes
are valid globally, i.e., over all effects and also over the formula to
regress. Note that this assumes that parameters have been reassigned
such that no two distinct parameter have the same name. Furthermore,
quantified effects and formulas must not use the same parameter names as
in ParameterTypes. This is not checked and must be guaranteed by the
caller.