regression: rework the regression of existential quantifiers
When regressing existential quantifiers, we need to consider all
alternative assignments. That means that we assign the variable with
all parameters of the action sequence and compute the resulting formula.
The resulting formula is the disjunction of all these alternatives.

Also rename the operator from 'some' to 'exists'.
Remove support for existential quantification in the form
exists(Var,Type,Formula). We only use variables lists from now on, i.e.:
exists([(Type,[Var1,Var2]),...],Formula)