generation: add first Python script that creates macro actions
The script defines a class MacroAction which will later contain all
properties of the macro. Currently, the class is mainly used to call
SWI-Prolog to create the macro, and the macro is then stored as a string
member of the class.

In main(), create a macro from the test domain.