generation: add a method generate to abstract from the method used
We currently have two ways to generate macro: We can use pyswip to
interface SWI-Prolog, or we use subprocess to call swipl directly. To
abstract from the method used, add a method generate that simply calls
one of the other methods (currently, we generate macros by starting
swipl in a subprocess).