upload script: use IDs instead of names to refer to domains, problems
In order to be able to make the distinction between augmented and
unaugmented domains with the same name, always refer to a domain by its
ID, not by its name. Also refer to problems by their ID for consistency
reasons.

If a domain name is given, look for the unaugmented domain in the
database and use that ID.