kubelet login controller: ping master to check if it's online
If the kubernetes master node is offline, the script fails because it
cannot reach the master. That failure usually takes a while because it
tries to connect several times. Also, this produces an unclear error
message at login. Thus, check first if the master if online and if not,
directly exit.