effects: constrain effect if partiall obliterated by subsequent effect
In a forall effect, it can happen that part of the effect is obliterated
by a subsequent effect, e.g. forall(o,p(o)) and not(p(o1)). In this
case, we need to adapt the effect to exclude all the conflicting parts,
e.g., forall(o,when(not(o=o1),p(o))).