effect tree: only consider conditional effects w/o vars bound by forall
If the condition of a conditional effect contains a variable which is
bound by a quantifier, we cannot make case distinctions but we must keep
the conditional effect as it is. This is because we don't know all
objects during macro generation, and therefore we cannot evaluate the
conditional, or even set it to true/false for individual objects.