effects: distinguish between quantified variables and parameters
For conditional effects in the previous effect, we need to distinguish
between parameters and quantified variables of the previous effect. This
is because for quantified variables, we need to bind the variables
again, while for parameters, we need to add the parameter to the current
effect.

Two examples are given in the tests.