effects: do not resolve a duplicated effect to nil, but keep it
The intention of reducing a duplicated effect (i.e., an effect that
already occurred before) was that we only need it once in the resulting
effect of the macro. However, in ADL, this effect can be nested in a
quantifier and may lead to unnecessary conditionals in the resulting
quantifier. As an example, forall(o,p(o)) with the previous effect p(a)
would have been resolved to forall(o,when(not(o=a)),p(o)). However,
simply keeping the full quantified effect forall(o,p(o)) is the desired
behavior here.

Duplicated effects are removed by the simplifier afterwards anyway, so
there is no reason to resolve them to nil.