macro evaluator: add evaluator that considers the complementarity
A set of macros has a high complementarity if the macros' actions
differ, i.e., if the macros do not contain the same actions. This is
useful because we do not want to add two very similar macros. The new
evaluator computes the complementarity of the given list of macros and
adds it as another factor to the evaluation of the weighed frequency
parameter reduction evaluator.