benchmark: also allow to filter by domain evaluator
This adds an option so the domains can be filtered by a domain
evaluator, not only by single macros. This allows to benchmark any
domains with a set of macros.

Also rename the option '--num-macros' to '--best' to reflect the change
above.

Set 'ff' as default planner if no planner is provided.