macro evaluator: divide the sum of macros by the sqrt of the macro count
Adding more macros should not always increase the total score of the set
of macros. But dividing by the number of macros would mean that the
singleton set consisting of the best macro would always be the
highest-scoring set, which is not desirable either. The square root of
the number of macros seems like a good middleground.