generation: add option to limit to most occurring action sequences
We may want to generate macros only for the most frequently occurring
action sequences. Add the option '--best n' which limits the macros to
the n most occurring action sequences. Note that this does not take
parameter assignment into account but uses the totalCount of the
sequence (thus only the most frequent assignment is relevant).