generation: update db entries if they exist
Instead of always creating new database entries for the generated macros
and augmented domains, check if the macro/domain already exists in the
database. If so, update the existing entry instead of adding a new one.
This way, we can re-run the generator without re-running all problems
for all domains.