effects: reformulate effects after formal definition
The formal definition of effect chaining provided some additional
insights, adapt the implementation accordingly.

Note that the current implementation does not work if a quantified
effect quantified a variable that occurs only in a nested condition of a
conditional effect.

Also, there are still a lot of choicepoints to be fixed.