regression of exists: do not backtrace over regress in findall
When we regress exists(...), we regress the inner formula over all
possible parameters. When we do this, we should regress each parameter
only once, otherwise wrong regression results are computed. We do this
simply by calling regress/4 instead of regress_/4, which also adds the
benefit that it already simplifies the results.

Also simplify the substituted formula before regressing it.
Additionally, fix a bug with the variable assignment in the result.