regression of exists: do not substitute with params
When regressing an existentially quantified formula, we do not need to
substitute with all known parameters, but instead we can just regress
with the quantified variable. This is because we already make the case
distinction if the variable is equal to another parameter or not. Thus,
substituting with known parameters/variables just adds the same
regressed formula twice.