identification: do not reset resulting actions in reduce step in loop
Set the actions of the reduced document once by using the actions of the
first document. We know that all documents have the same actions
(because the key is computed with the actions), therefore there is no
need to collect the actions in any way. The loop was supposed to work on
the parameter assignments only.

This fixes a bug where the action sequence sometimes contained a NaN
action at the end.