stats: collect descriptives in a dict, also support PTT
We want to collect all descriptives so we can insert them into a latex
template. Therefore, do not only print the descriptives, but also
collect them.

Additionally, if the planner is 'ptt', add the suffix '_ptt' to the
domain name and use ff as the actual planner. This is because ptt also
generates augmented domains, which is uploaded separately and assumed to
have the suffix '_ptt'.