planning worker: also caption output of FdSatPlanner
We do not start the planner with subprocess.run but use Popen instead,
because we need to run it in the background and stop it if we found a
plan. Therefore, add a method get_output() to the abstract Planner class
and add implementations of the method for all planners.