stats: change descriptives format to better deal with multiple configs
We have a lot of possible dbmp configurations and we may want to get the
descriptives for only the best n configurations. Add an option to select
those. Also change the represenation of the descriptives, so the
configuration is no longer part of the planner name but instead appears
as a separate config field. As we can now have multiple entries for the
same planner, use a list instead of a dict.