generator: add option to store only best n macro domains
So far we could only select individual macros, but we couldn't select
macro domains to store. Add an option '--store-best n' that stores an
augmented domain if it is one of the best n domains according to any
evaluator. Note that this may store more domains, as we only compare to
the best n augmented domains that we have evaluated so far. This still
significantly reduces the number of stored domains and speeds up the
generation process.